refactor(llm): extract provider seam — base/orchestration/anthropic split (#274)
Splits app/llm/client.py (742 lines, Anthropic-only) into a Protocol-based
provider interface, neutral order/tool helpers, and an AnthropicLLM
implementation. Adds get_llm() factory mirroring the STT/TTS pattern in
app/stt/__init__.py and app/tts/__init__.py.

  app/llm/
    __init__.py        # get_llm() factory + re-exports
    base.py            # LLMProvider Protocol, StreamEvent, LLMResponse,
                       # ToolSpec, UPDATE_ORDER_TOOL_SPEC (neutral)
    orchestration.py   # apply_order_patch, validate_order_patch,
                       # summarize_order_for_tool_result, should_skip_followup
    anthropic.py       # AnthropicLLM class — Anthropic-specific wire format

Surfaces ANTHROPIC_MODEL and ANTHROPIC_MAX_TOKENS as config so swapping
Haiku → Sonnet → Opus is an env change, not a code edit. LLM_PROVIDER /
STT_PROVIDER / TTS_PROVIDER documented in .env.example.

Documented two known leaks in base.py:
- StreamEvent.timing payload still carries Anthropic-flavored fields
  (cache_read/creation_tokens, network_prefill/decode_seconds); other
  providers should report 0 for fields they cannot derive.
- LLMResponse.history is opaque to the router but provider-private in
  shape — no hot-swap mid-call.

No behavior change — same Anthropic SDK calls, same tool schema, same
prompt-cache breakpoints, same first-audio latency path. 488 backend
tests pass (2 pre-existing recordings-storage failures unrelated).

Closes #274
