+++ .env.example
@@ -4,6 +4,23 @@
 # Anthropic — Claude Haiku 4.5
 ANTHROPIC_API_KEY=
 
+# LLM provider. Today only "anthropic" is implemented; the selector
+# seam in app/llm/__init__.py is ready for OpenAI/Gemini if/when
+# we add them. Leave unset to use the default.
+LLM_PROVIDER=anthropic
+
+# Anthropic model + reply cap. Defaults are Haiku 4.5 / 512 tokens.
+# Swap to claude-sonnet-4-6 or claude-opus-4-7 for evaluation —
+# expect ~2-3× TTFT and ~5× cost on Sonnet vs Haiku.
+ANTHROPIC_MODEL=claude-haiku-4-5-20251001
+ANTHROPIC_MAX_TOKENS=512
+
+# STT + TTS provider selectors. Today only "deepgram" is implemented
+# for each; the seams in app/stt/__init__.py and app/tts/__init__.py
+# are ready for a second vendor (e.g. Whisper STT, ElevenLabs TTS).
+STT_PROVIDER=deepgram
+TTS_PROVIDER=deepgram
+
 # Deepgram — streaming STT (Flux English mono) and TTS (Aura)
 DEEPGRAM_API_KEY=
 DEEPGRAM_TTS_MODEL=aura-2-thalia-en
