--- app/llm/client.py
@@ -1,741 +1,0 @@
-"""Anthropic Claude Haiku 4.5 client for the voice ordering agent.
-
-Takes a caller transcript plus conversation history and the current
-order state, returns the assistant's spoken reply along with any
-updates to the order (produced via the ``update_order`` tool).
-
-The tool schema mirrors ``app.orders.models.Order`` so Haiku can emit
-partial order state incrementally. The model can both speak to the
-caller and call ``update_order`` in the same turn — we process both
-and advance the conversation.
-
-Two entry points:
-
-- ``generate_reply`` — synchronous round-trip. Used by tests and any
-  offline / batch path. Waits for Haiku to finish before returning.
-- ``stream_reply`` — async generator that yields text deltas as Haiku
-  produces them, then yields a terminal event with the final order
-  and threaded history. Used by the call-flow orchestrator (#40) to
-  start TTS before the full reply is ready and hit the <1s first-audio
-  latency budget.
-"""
-
-from __future__ import annotations
-
-import time
-from dataclasses import dataclass
-from typing import Any, AsyncIterator, Optional
-
-from anthropic import Anthropic, AsyncAnthropic
-
-from app.config import settings
-from app.orders.models import Order
-from app.orders.validation import validate_delivery_address
-
-MODEL = "claude-haiku-4-5-20251001"
-MAX_TOKENS = 512
-
-UPDATE_ORDER_TOOL: dict[str, Any] = {
-    "name": "update_order",
-    "description": (
-        "Record the caller's current order state. Call this whenever "
-        "the order changes — items added, removed, or modified; order "
-        "type decided; delivery address given; or the caller confirms "
-        "or cancels. Emit the FULL current order state each time, not "
-        "a diff."
-    ),
-    "input_schema": {
-        "type": "object",
-        "properties": {
-            "items": {
-                "type": "array",
-                "description": "Full list of line items currently in the order.",
-                "items": {
-                    "type": "object",
-                    "properties": {
-                        "name": {"type": "string"},
-                        "category": {
-                            "type": "string",
-                            "description": (
-                                "Free-form category from the menu (e.g. "
-                                "appetizer, main, soup, drink, dessert). "
-                                "Used for grouping in the dashboard — not "
-                                "validated against a fixed enum, since "
-                                "tenants pick their own category names."
-                            ),
-                        },
-                        "size": {
-                            "type": ["string", "null"],
-                            "description": (
-                                "Required when the menu item is multi-size "
-                                "(small/medium/large, half/whole, etc.) — "
-                                "use whichever size keys the menu shows. "
-                                "Null for single-priced items."
-                            ),
-                        },
-                        "quantity": {"type": "integer", "minimum": 1},
-                        "unit_price": {
-                            "type": "number",
-                            "description": "Per-unit price from the menu.",
-                        },
-                        "modifications": {
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "description": ("Customizations like 'extra cheese' or 'no onions'."),
-                        },
-                    },
-                    "required": ["name", "category", "quantity", "unit_price"],
-                },
-            },
-            "order_type": {
-                "type": ["string", "null"],
-                "enum": ["pickup", "delivery", None],
-            },
-            "delivery_address": {"type": ["string", "null"]},
-            "status": {
-                "type": "string",
-                "enum": ["in_progress", "confirmed", "cancelled"],
-            },
-        },
-    },
-}
-
-
-@dataclass
-class LLMResponse:
-    reply_text: str
-    order: Order
-    history: list[dict[str, Any]]
-
-
-@dataclass
-class StreamEvent:
-    """One event yielded by ``stream_reply``.
-
-    Exactly one of ``text_delta``, ``timing``, or ``final`` is set on
-    any given event. Text-delta events arrive incrementally as Haiku
-    produces output; a single ``timing`` event lands the moment the
-    first text block opens (so the router can fold the latency
-    breakdown into its ``first_audio`` Firestore event before TTS
-    starts); the terminal ``final`` event carries the assembled
-    ``LLMResponse`` so the caller can persist state for the next turn.
-
-    The ``timing`` payload has the shape::
-
-        {
-            "ttft_seconds": float,             # request → first SDK event
-            "tool_prefix_seconds": float,      # first event → first text block
-            "network_prefill_seconds": float,  # request → message_start (#175)
-            "decode_seconds": float,           # message_start → first text block (#175)
-            "cache_read_tokens": int,
-            "cache_creation_tokens": int,
-        }
-
-    On tool-use turns a second timing event is emitted for the follow-up
-    call so each network round-trip is measured independently (#175).
-    """
-
-    text_delta: Optional[str] = None
-    final: Optional[LLMResponse] = None
-    timing: Optional[dict[str, Any]] = None
-
-
-def _missing_key_error() -> RuntimeError:
-    return RuntimeError(
-        "ANTHROPIC_API_KEY not set — cannot call the LLM. "
-        "Populate it in your .env (fetch via /shared-creds)."
-    )
-
-
-def _client() -> Anthropic:
-    key = settings.anthropic_api_key
-    if not key:
-        raise _missing_key_error()
-    # Sync path is only used by tests; intentionally not a singleton so tests
-    # can patch/inject freely without worrying about module-level state. If this
-    # ever needs to be a singleton too, follow the same lazy pattern as
-    # _get_async_client below — but don't change it speculatively.
-    return Anthropic(api_key=key)
-
-
-# Process-wide reusable AsyncAnthropic instance (#175). Constructing a new
-# AsyncAnthropic() on every stream_reply call creates a fresh httpx.AsyncClient
-# per turn; when the local reference goes out of scope the pool closes and the
-# TLS connection is torn down, costing a full TCP+TLS handshake on the next
-# turn (~100-250ms). Reusing one instance keeps the connection pool warm.
-# Lazy-initialised so importing this module never spins up sockets at startup.
-# Reset to None in tests that need a clean slate (see _reset_async_client).
-_default_async_client: Optional[AsyncAnthropic] = None
-
-
-def _get_async_client() -> AsyncAnthropic:
-    global _default_async_client
-    if _default_async_client is None:
-        key = settings.anthropic_api_key
-        if not key:
-            raise _missing_key_error()
-        _default_async_client = AsyncAnthropic(api_key=key)
-    return _default_async_client
-
-
-def _reset_async_client() -> None:
-    """Reset the module-level singleton. Called by tests that inject their own
-    fake client so they don't accidentally touch the real singleton."""
-    global _default_async_client
-    _default_async_client = None
-
-
-def _serialize_block(block: Any) -> dict[str, Any]:
-    """Convert an SDK content block to the dict shape the Messages API accepts.
-
-    The Anthropic SDK's streaming blocks carry extra attributes like
-    ``parsed_output`` that ``model_dump()`` faithfully serializes — but the
-    Messages API rejects unknown fields, so any subsequent turn that threads
-    that history 400s. We rebuild the API-valid shape by hand instead.
-    """
-    if block.type == "text":
-        return {"type": "text", "text": block.text}
-    if block.type == "tool_use":
-        return {
-            "type": "tool_use",
-            "id": block.id,
-            "name": block.name,
-            "input": block.input,
-        }
-    raise ValueError(f"Unsupported content block type: {block.type!r}")
-
-
-def _summarize_order(order: Order) -> str:
-    """Server-side summary fed back to the LLM as a ``tool_result``.
-
-    Returning the post-apply subtotal + item list closes the
-    accuracy hole observed on the 2026-04-26 Twilight call where
-    Haiku quoted a $50.50 total for an order that actually summed to
-    $49.25 — without a server-verified number in the tool_result, the
-    model fabricates totals from its memory of unit prices. Now it
-    has a ground-truth subtotal to read back instead.
-    """
-    if not order.items:
-        return "Order updated. Subtotal: $0.00. (no items yet)"
-    items_summary = ", ".join(
-        f"{item.quantity}× {item.name}"
-        + (f" ({', '.join(item.modifications)})" if item.modifications else "")
-        for item in order.items
-    )
-    return f"Order updated. Subtotal: ${order.subtotal:.2f}. Items: {items_summary}."
-
-
-def _tool_result_block(tool_use_id: str, content: str = "Order updated.") -> dict[str, Any]:
-    return {
-        "type": "tool_result",
-        "tool_use_id": tool_use_id,
-        "content": content,
-    }
-
-
-def _system_cache_block(system_prompt: str) -> list[dict[str, Any]]:
-    """Wrap the system prompt for Anthropic prompt caching.
-
-    Passing system as a list with cache_control=ephemeral tells Anthropic
-    to cache the block server-side for up to 5 minutes. Cache hits reduce
-    system-prompt token cost by ~90% and shave first-token latency on
-    turns 2+ of the same call — near-certain for any call lasting > 20s.
-    """
-    return [
-        {
-            "type": "text",
-            "text": system_prompt,
-            "cache_control": {"type": "ephemeral"},
-        }
-    ]
-
-
-def _with_rolling_cache_breakpoint(
-    messages: list[dict[str, Any]],
-) -> list[dict[str, Any]]:
-    """Mark the last block of ``messages[-1]`` with ``cache_control: ephemeral``.
-
-    Together with the system-prompt breakpoint, this gives a 2-breakpoint
-    rolling cache (#176): each turn's request is cached up through the
-    new last message, and the next turn reads that as its cached prefix
-    so only the new content (previous assistant reply + new user turn)
-    is uncached prefill.
-
-    Caveat: Haiku 4.5 has a 4096-token minimum cacheable prefix. Short
-    calls won't engage the rolling breakpoint; longer calls (turn 3+ of
-    a typical order) will.
-
-    Returns a NEW messages list with a NEW last message — the input is
-    not mutated. This matters because callers thread the original
-    history forward into ``LLMResponse.history``; a stale cache_control
-    marker in threaded history would accumulate every turn and quickly
-    blow past Anthropic's 4-breakpoint cap.
-    """
-    if not messages:
-        return messages
-    last = messages[-1]
-    content = last["content"]
-    if isinstance(content, str):
-        # Plain-string user transcript. Convert to a one-block list so we
-        # can attach cache_control — strings have nowhere to attach it.
-        new_content: list[dict[str, Any]] = [
-            {
-                "type": "text",
-                "text": content,
-                "cache_control": {"type": "ephemeral"},
-            }
-        ]
-    else:
-        if not content:
-            return messages  # defensive: nothing to mark
-        new_content = [dict(block) for block in content]
-        new_content[-1] = {
-            **new_content[-1],
-            "cache_control": {"type": "ephemeral"},
-        }
-    return [*messages[:-1], {**last, "content": new_content}]
-
-
-def _append_user_transcript(history: list[dict[str, Any]], transcript: str) -> list[dict[str, Any]]:
-    """Append the caller's transcript to history with valid alternation.
-
-    Anthropic requires strict user/assistant alternation. After a turn
-    that produced both text and a ``tool_use``, history ends in a synthetic
-    ``user: [tool_result]`` message. Appending another ``user`` message
-    would error, so we merge the new transcript into that pending
-    ``tool_result`` message instead.
-    """
-    if (
-        history
-        and history[-1]["role"] == "user"
-        and isinstance(history[-1]["content"], list)
-        and history[-1]["content"]
-        and history[-1]["content"][0].get("type") == "tool_result"
-    ):
-        merged = [
-            *history[-1]["content"],
-            {"type": "text", "text": transcript},
-        ]
-        return [*history[:-1], {"role": "user", "content": merged}]
-    return [*history, {"role": "user", "content": transcript}]
-
-
-_INVALID_ADDRESS_NOTE = (
-    "Delivery address incomplete — please ask the caller for the full street address."
-)
-
-
-def _apply_validation(patch: dict[str, Any]) -> tuple[dict[str, Any], list[str]]:
-    """Filter an update_order patch through server-side validators.
-
-    Returns a (cleaned_patch, rejection_notes) tuple:
-    - cleaned_patch is a copy of patch with any field that failed
-      validation removed (so the previous Order value stays put when
-      the patch is applied).
-    - rejection_notes is a list of human-readable strings to append to
-      the tool_result so Haiku knows to re-ask the caller. Empty when
-      every field passed validation.
-
-    Today only delivery_address has a validator (Sprint 2.2 #105). New
-    field validators slot in here so _apply_update stays a dumb
-    dict-merger and orchestration stays in one place.
-
-    Note on explicit-clear intents: when Haiku ships
-    delivery_address=None or "" (e.g. swapping order_type from
-    delivery to pickup), that's a legitimate clear, not a rejection.
-    The validator returns False for both, so we only invoke it when
-    there's actual non-empty content to validate.
-    """
-    cleaned = dict(patch)
-    notes: list[str] = []
-    if "delivery_address" in cleaned:
-        value = cleaned["delivery_address"]
-        # None / empty / whitespace-only is an explicit clear — pass through.
-        # Only validate when the caller actually provided content.
-        if value is not None and isinstance(value, str) and value.strip():
-            if not validate_delivery_address(value):
-                del cleaned["delivery_address"]
-                notes.append(_INVALID_ADDRESS_NOTE)
-    return cleaned, notes
-
-
-_TERMINAL_STATUSES = frozenset({"confirmed", "cancelled"})
-
-
-def _should_skip_followup(tool_uses: list[dict[str, Any]], main_emitted_text: bool) -> bool:
-    """Decide whether to skip the post-tool-use follow-up LLM call (#270).
-
-    The follow-up exists to give the model a chance to speak after seeing
-    ``tool_result`` feedback (#173). On terminal-status turns (confirmed
-    or cancelled) the call is ending — there is no next caller turn for
-    the model to react to — and if the main call already emitted text the
-    caller has already heard the goodbye. In that case the follow-up is
-    pure latency.
-
-    The follow-up still runs when the main call was tool-only on a
-    terminal turn (caller hasn't heard anything yet), or on any
-    non-terminal turn (mid-order item adds need the server-verified
-    subtotal in tool_result before the next "anything else?").
-    """
-    if not main_emitted_text:
-        return False
-    return any(tu["input"].get("status") in _TERMINAL_STATUSES for tu in tool_uses)
-
-
-def _apply_update(order: Order, patch: dict[str, Any]) -> Order:
-    """Merge a tool-call payload into the current Order.
-
-    Preserves call_sid, caller_phone, restaurant_id, and created_at
-    from the existing order so the LLM cannot overwrite them. Every
-    other field the LLM provided is accepted as the new authoritative
-    value.
-    """
-
-    preserved = {
-        "call_sid": order.call_sid,
-        "caller_phone": order.caller_phone,
-        "restaurant_id": order.restaurant_id,
-        "created_at": order.created_at,
-    }
-    merged = {**order.model_dump(), **patch, **preserved}
-    return Order.model_validate(merged)
-
-
-def generate_reply(
-    *,
-    transcript: str,
-    history: list[dict[str, Any]],
-    order: Order,
-    system_prompt: str,
-    client: Optional[Anthropic] = None,
-) -> LLMResponse:
-    """Send the caller's latest transcript to Haiku and return a reply.
-
-    ``system_prompt`` is rendered per call from the inbound restaurant's
-    config (#79); the previously-cached module-level prompt is gone.
-
-    ``history`` is Anthropic's Messages format: a list of
-    ``{"role": "user"|"assistant", "content": ...}`` dicts. The updated
-    history (with the new user turn and the assistant turn appended) is
-    returned in ``LLMResponse.history`` so the caller can thread it
-    into the next turn verbatim — including any tool_use blocks, which
-    Anthropic requires to stay in context.
-    """
-
-    api = client or _client()
-
-    new_history = _append_user_transcript(history, transcript)
-
-    response = api.messages.create(
-        model=MODEL,
-        max_tokens=MAX_TOKENS,
-        system=_system_cache_block(system_prompt),
-        tools=[UPDATE_ORDER_TOOL],
-        messages=_with_rolling_cache_breakpoint(new_history),
-    )
-
-    reply_text_parts: list[str] = []
-    tool_uses: list[dict[str, Any]] = []
-    for block in response.content:
-        if block.type == "text":
-            reply_text_parts.append(block.text)
-        elif block.type == "tool_use" and block.name == "update_order":
-            tool_uses.append({"id": block.id, "input": block.input})
-
-    updated_order = order
-    tool_results: list[dict[str, Any]] = []
-    for tu in tool_uses:
-        cleaned_input, rejection_notes = _apply_validation(tu["input"])
-        updated_order = _apply_update(updated_order, cleaned_input)
-        summary = _summarize_order(updated_order)
-        if rejection_notes:
-            summary = summary + " " + " ".join(rejection_notes)
-        tool_results.append(_tool_result_block(tu["id"], summary))
-
-    assistant_content = [_serialize_block(block) for block in response.content]
-    new_history = [
-        *new_history,
-        {"role": "assistant", "content": assistant_content},
-    ]
-
-    if tool_uses:
-        new_history = [
-            *new_history,
-            {"role": "user", "content": tool_results},
-        ]
-
-    if tool_uses and not _should_skip_followup(tool_uses, bool(reply_text_parts)):
-        # Verbal continuation only — no further tool calls (#173). We keep the
-        # tool schema in ``tools`` and use ``tool_choice="none"`` to suppress
-        # tool use, rather than passing ``tools=[]``: tools sit before system
-        # in Anthropic's cached prefix order, so changing them across the two
-        # calls of one turn invalidates the entire cached prefix on the
-        # follow-up (system + menu + history). See #176.
-        followup = api.messages.create(
-            model=MODEL,
-            max_tokens=MAX_TOKENS,
-            system=_system_cache_block(system_prompt),
-            tools=[UPDATE_ORDER_TOOL],
-            tool_choice={"type": "none"},
-            messages=_with_rolling_cache_breakpoint(new_history),
-        )
-        for block in followup.content:
-            if block.type == "text":
-                reply_text_parts.append(block.text)
-        followup_content = [_serialize_block(block) for block in followup.content]
-        new_history = [
-            *new_history,
-            {"role": "assistant", "content": followup_content},
-        ]
-
-    return LLMResponse(
-        reply_text="".join(reply_text_parts).strip(),
-        order=updated_order,
-        history=new_history,
-    )
-
-
-async def stream_reply(
-    *,
-    transcript: str,
-    history: list[dict[str, Any]],
-    order: Order,
-    system_prompt: str,
-    client: Optional[AsyncAnthropic] = None,
-) -> AsyncIterator[StreamEvent]:
-    """Stream Haiku's reply token-by-token for low-latency TTS handoff.
-
-    Yields ``StreamEvent(text_delta=str)`` for each incremental chunk
-    of the assistant's spoken reply, then a single terminal
-    ``StreamEvent(final=LLMResponse)`` with the full reply text,
-    updated order, and threaded history.
-
-    The contract matches ``generate_reply`` for tool-use semantics:
-    if the first turn emits only ``update_order`` blocks (no text),
-    we send a ``tool_result`` and stream a follow-up call so the
-    caller still gets a spoken reply. In practice this is the
-    "I cancelled the order" path — the model often wants to confirm
-    the side effect before talking back.
-    """
-
-    api = client or _get_async_client()
-
-    new_history = _append_user_transcript(history, transcript)
-
-    text_parts: list[str] = []
-    tool_uses: list[dict[str, Any]] = []
-
-    # Latency instrumentation (#146, extended in #175). t_request_start anchors
-    # the whole turn — TTFT is time to first SDK event; tool_prefix is the gap
-    # between first event and first text block. #175 adds two finer fields:
-    # - network_prefill_seconds: request → message_start (TCP+TLS+queue+prefill)
-    # - decode_seconds: message_start → first text block (token-1 decode)
-    t_request_start = time.monotonic()
-    t_first_event: Optional[float] = None
-    t_message_start: Optional[float] = None
-    t_first_text_block: Optional[float] = None
-    cache_read_tokens = 0
-    cache_creation_tokens = 0
-    timing_emitted = False
-
-    def _make_timing_event(
-        _t_req: float,
-        _t_first_ev: Optional[float],
-        _t_msg_start: Optional[float],
-        _t_first_text: Optional[float],
-        _cache_read: int,
-        _cache_creation: int,
-    ) -> StreamEvent:
-        return StreamEvent(
-            timing={
-                "ttft_seconds": round((_t_first_ev or _t_req) - _t_req, 3),
-                "tool_prefix_seconds": round(
-                    (_t_first_text - _t_first_ev) if (_t_first_text and _t_first_ev) else 0.0,
-                    3,
-                ),
-                "network_prefill_seconds": round(
-                    (_t_msg_start - _t_req) if _t_msg_start is not None else 0.0,
-                    3,
-                ),
-                "decode_seconds": round(
-                    (_t_first_text - _t_msg_start)
-                    if (_t_first_text is not None and _t_msg_start is not None)
-                    else 0.0,
-                    3,
-                ),
-                "cache_read_tokens": _cache_read,
-                "cache_creation_tokens": _cache_creation,
-            }
-        )
-
-    async with api.messages.stream(
-        model=MODEL,
-        max_tokens=MAX_TOKENS,
-        system=_system_cache_block(system_prompt),
-        tools=[UPDATE_ORDER_TOOL],
-        messages=_with_rolling_cache_breakpoint(new_history),
-    ) as stream:
-        async for event in stream:
-            if t_first_event is None:
-                t_first_event = time.monotonic()
-            etype = getattr(event, "type", None)
-            if etype == "message_start":
-                t_message_start = time.monotonic()
-                msg = getattr(event, "message", None)
-                usage = getattr(msg, "usage", None) if msg is not None else None
-                if usage is not None:
-                    cache_read_tokens = getattr(usage, "cache_read_input_tokens", 0) or 0
-                    cache_creation_tokens = getattr(usage, "cache_creation_input_tokens", 0) or 0
-            elif etype == "content_block_start":
-                block = getattr(event, "content_block", None)
-                if getattr(block, "type", None) == "text" and t_first_text_block is None:
-                    t_first_text_block = time.monotonic()
-                    yield _make_timing_event(
-                        t_request_start,
-                        t_first_event,
-                        t_message_start,
-                        t_first_text_block,
-                        cache_read_tokens,
-                        cache_creation_tokens,
-                    )
-                    timing_emitted = True
-            elif etype == "content_block_delta":
-                delta = getattr(event, "delta", None)
-                if getattr(delta, "type", None) == "text_delta":
-                    text_parts.append(delta.text)
-                    yield StreamEvent(text_delta=delta.text)
-        first_message = await stream.get_final_message()
-
-    for block in first_message.content:
-        if block.type == "tool_use" and block.name == "update_order":
-            tool_uses.append({"id": block.id, "input": block.input})
-
-    updated_order = order
-    tool_results: list[dict[str, Any]] = []
-    for tu in tool_uses:
-        cleaned_input, rejection_notes = _apply_validation(tu["input"])
-        updated_order = _apply_update(updated_order, cleaned_input)
-        summary = _summarize_order(updated_order)
-        if rejection_notes:
-            summary = summary + " " + " ".join(rejection_notes)
-        tool_results.append(_tool_result_block(tu["id"], summary))
-
-    assistant_content = [_serialize_block(block) for block in first_message.content]
-    new_history = [
-        *new_history,
-        {"role": "assistant", "content": assistant_content},
-    ]
-
-    if tool_uses:
-        new_history = [
-            *new_history,
-            {"role": "user", "content": tool_results},
-        ]
-
-    if tool_uses and not timing_emitted:
-        # First stream produced only tool_use blocks (no text). Emit the
-        # first-call timing snapshot before starting the follow-up so the
-        # router always receives a timing event per network round-trip (#175).
-        yield _make_timing_event(
-            t_request_start,
-            t_first_event,
-            t_message_start,
-            t_first_text_block,
-            cache_read_tokens,
-            cache_creation_tokens,
-        )
-
-    if tool_uses and not _should_skip_followup(tool_uses, bool(text_parts)):
-        # Verbal continuation only — no further tool calls (#173). We keep the
-        # tool schema in ``tools`` and use ``tool_choice="none"`` to suppress
-        # tool use, rather than passing ``tools=[]``: tools sit before system
-        # in Anthropic's cached prefix order, so changing them across the two
-        # calls of one turn invalidates the entire cached prefix on the
-        # follow-up (system + menu + history). See #176.
-        # Each follow-up call gets its own scoped timing variables so the second
-        # timing event reports that call's numbers, not the first call's (#175).
-        fu_t_request_start = time.monotonic()
-        fu_t_first_event: Optional[float] = None
-        fu_t_message_start: Optional[float] = None
-        fu_t_first_text_block: Optional[float] = None
-        fu_cache_read_tokens = 0
-        fu_cache_creation_tokens = 0
-        fu_timing_emitted = False
-
-        async with api.messages.stream(
-            model=MODEL,
-            max_tokens=MAX_TOKENS,
-            system=_system_cache_block(system_prompt),
-            tools=[UPDATE_ORDER_TOOL],
-            tool_choice={"type": "none"},
-            messages=_with_rolling_cache_breakpoint(new_history),
-        ) as followup_stream:
-            async for event in followup_stream:
-                if fu_t_first_event is None:
-                    fu_t_first_event = time.monotonic()
-                etype = getattr(event, "type", None)
-                if etype == "message_start":
-                    fu_t_message_start = time.monotonic()
-                    msg = getattr(event, "message", None)
-                    usage = getattr(msg, "usage", None) if msg is not None else None
-                    if usage is not None:
-                        fu_cache_read_tokens = getattr(usage, "cache_read_input_tokens", 0) or 0
-                        fu_cache_creation_tokens = (
-                            getattr(usage, "cache_creation_input_tokens", 0) or 0
-                        )
-                elif etype == "content_block_start":
-                    block = getattr(event, "content_block", None)
-                    if getattr(block, "type", None) == "text" and fu_t_first_text_block is None:
-                        fu_t_first_text_block = time.monotonic()
-                        yield _make_timing_event(
-                            fu_t_request_start,
-                            fu_t_first_event,
-                            fu_t_message_start,
-                            fu_t_first_text_block,
-                            fu_cache_read_tokens,
-                            fu_cache_creation_tokens,
-                        )
-                        fu_timing_emitted = True
-                elif etype == "content_block_delta":
-                    delta = getattr(event, "delta", None)
-                    if getattr(delta, "type", None) == "text_delta":
-                        text_parts.append(delta.text)
-                        yield StreamEvent(text_delta=delta.text)
-            followup_message = await followup_stream.get_final_message()
-
-        if not fu_timing_emitted:
-            # Follow-up emitted no text blocks; emit a timing snapshot anyway.
-            yield _make_timing_event(
-                fu_t_request_start,
-                fu_t_first_event,
-                fu_t_message_start,
-                fu_t_first_text_block,
-                fu_cache_read_tokens,
-                fu_cache_creation_tokens,
-            )
-
-        followup_content = [_serialize_block(b) for b in followup_message.content]
-        new_history = [
-            *new_history,
-            {"role": "assistant", "content": followup_content},
-        ]
-
-    if not timing_emitted and not tool_uses:
-        # Tool-only path with no follow-up text (rare); still emit a
-        # timing snapshot so consumers always see the breakdown.
-        yield _make_timing_event(
-            t_request_start,
-            t_first_event,
-            t_message_start,
-            t_first_text_block,
-            cache_read_tokens,
-            cache_creation_tokens,
-        )
-
-    yield StreamEvent(
-        final=LLMResponse(
-            reply_text="".join(text_parts).strip(),
-            order=updated_order,
-            history=new_history,
-        )
-    )
